+++ HISTORY.md
@@ -6,6 +6,9 @@
 - Support for mount filesystems
 - Check staging and publishing targets
 - Exit on binding error
+- Check presence of REQUIRED fields
+- Log requests and responses
+- Support defining default mount filesystem
 
 
 ## 0.0.2 (2018-06-19)